requirements.txt
blinker==1.9.0
# bzip2==1.0.8
# ca-certificates==2025.1.31
click==8.1.8
colorama==0.4.6
flask==3.1.0
importlib-metadata==8.6.1
# intel-openmp==2025.0.0
itsdangerous==2.2.0
jinja2==3.1.5
# libblas==3.9.0
# libcblas==3.9.0
# libexpat==2.6.4
# libffi==3.4.6
# liblapack==3.9.0
# liblzma==5.6.4
# libsqlite==3.48.0
# libzlib==1.3.1
markupsafe==3.0.2
# mkl==2020.4
numpy==2.2.3
# openssl==3.4.1
pandas==2.2.3
# pip==25.0.1
# python==3.12.9
python-dateutil==2.9.0.post0
# python-tzdata==2025.1
# python_abi==3.12
pytz==2024.1
setuptools==75.8.0
six==1.17.0
# tk==8.6.13
tzdata==2024.2
# ucrt==10.0.22621.0
# vc==14.3
# vc14_runtime==14.42.34433
werkzeug==3.1.3
wheel==0.45.1
zipp==3.21.0
azure-storage-blob==12.20.0